refactor: rename module to reflect method name
